Reorg of includes/class names
