+++ src/InterpreterContext.h
@@ -1,0 +1,81 @@
+/*
+ * Souffle - A Datalog Compiler
+ * Copyright (c) 2013, 2015, Oracle and/or its affiliates. All rights reserved
+ * Licensed under the Universal Permissive License v 1.0 as shown at:
+ * - https://opensource.org/licenses/UPL
+ * - <souffle root>/licenses/SOUFFLE-UPL.txt
+ */
+
+/************************************************************************
+ *
+ * @file InterpreterContext.h
+ *
+ * Defines interpreter context 
+ *
+ ***********************************************************************/
+
+#pragma once
+
+#include "RamTypes.h"
+#include <vector>
+
+namespace souffle {
+
+/**
+ * Evaluation context for RAM operations
+ */
+class InterpreterContext {
+    std::vector<const RamDomain*> data;
+    std::vector<RamDomain>* returnValues = nullptr;
+    std::vector<bool>* returnErrors = nullptr;
+    const std::vector<RamDomain>* args = nullptr;
+
+public:
+    InterpreterContext(size_t size = 0) : data(size) {}
+    virtual ~InterpreterContext() = default;
+
+    const RamDomain*& operator[](size_t index) {
+        return data[index];
+    }
+
+    const RamDomain* const& operator[](size_t index) const {
+        return data[index];
+    }
+
+    std::vector<RamDomain>& getReturnValues() const {
+        return *returnValues;
+    }
+
+    void setReturnValues(std::vector<RamDomain>& retVals) {
+        returnValues = &retVals;
+    }
+
+    void addReturnValue(RamDomain val, bool err = false) {
+        assert(returnValues != nullptr && returnErrors != nullptr);
+        returnValues->push_back(val);
+        returnErrors->push_back(err);
+    }
+
+    std::vector<bool>& getReturnErrors() const {
+        return *returnErrors;
+    }
+
+    void setReturnErrors(std::vector<bool>& retErrs) {
+        returnErrors = &retErrs;
+    }
+
+    const std::vector<RamDomain>& getArguments() const {
+        return *args;
+    }
+
+    void setArguments(const std::vector<RamDomain>& a) {
+        args = &a;
+    }
+
+    RamDomain getArgument(size_t i) const {
+        assert(args != nullptr && i < args->size() && "argument out of range");
+        return (*args)[i];
+    }
+};
+
+}  // end of namespace souffle
